obsidian_plugin_backend/requirements.txt
accelerate==1.10.1
acres==0.5.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.10.0
astor==0.8.1
asyncpg==0.30.0
attrs==25.3.0
bcrypt==4.3.0
beautifulsoup4==4.13.5
blake3==1.0.5
cachetools==6.2.0
cbor2==5.7.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
chz==0.3.0
ci-info==0.3.0
click==8.2.1
cloudpickle==3.1.1
compressed-tensors==0.10.2
configobj==5.0.9
configparser==7.2.0
contourpy==1.3.3
cuda-bindings==13.0.1
cuda-pathfinder==1.2.2
cuda-python==13.0.1
cupy-cuda12x==13.6.0
cycler==0.12.1
depyf==0.19.0
dill==0.4.0
diskcache==5.6.3
distro==1.9.0
dnspython==2.8.0
docker==7.1.0
docling==2.53.0
docling-core==2.48.1
docling-ibm-models==3.9.1
docling-parse==4.5.0
easyocr==1.7.2
einops==0.8.1
email-validator==2.3.0
et_xmlfile==2.0.0
etelemetry==0.3.1
Faker==37.8.0
fastapi==0.116.1
fastapi-cli==0.0.11
fastapi-cloud-cli==0.1.5
fastrlock==0.8.3
filelock==3.19.1
filetype==1.2.0
flashinfer-python==0.2.8
fonttools==4.60.1
frozenlist==1.7.0
fsspec==2025.9.0
gguf==0.17.1
gpt-oss==0.1.0
h11==0.16.0
hf-xet==1.1.10
html2text==2025.4.15
httpcore==1.0.9
httplib2==0.31.0
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.35.0rc1
idna==3.10
imageio==2.37.0
importlib_metadata==8.7.0
iniconfig==2.1.0
interegular==0.3.3
Jinja2==3.1.6
jiter==0.11.0
jsonlines==3.1.0
jsonref==1.1.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
lark==1.2.2
latex2mathml==3.78.1
lazy_loader==0.4
llguidance==0.7.30
llvmlite==0.44.0
lm-format-enforcer==0.10.12
looseversion==1.3.0
lxml==5.4.0
markdown-it-py==4.0.0
marko==2.2.0
MarkupSafe==3.0.2
matplotlib==3.10.7
mcp==1.14.0
mdurl==0.1.2
mistral_common==1.8.5
mpire==2.10.2
mpmath==1.3.0
msgpack==1.1.1
msgspec==0.19.0
multidict==6.6.4
multiprocess==0.70.18
networkx==3.5
nibabel==5.3.2
ninja==1.13.0
nipype==1.10.0
numba==0.61.2
numpy==2.2.6
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cudnn-frontend==1.14.1
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-ml-py==13.580.82
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.9
nvidia-nvtx-cu12==12.8.90
openai==1.107.3
openai-harmony==0.0.4
opencv-python-headless==4.12.0.88
openpyxl==3.1.5
opentelemetry-api==1.37.0
opentelemetry-sdk==1.37.0
opentelemetry-semantic-conventions==0.58b0
outlines_core==0.2.10
packaging==25.0
pandas==2.3.2
partial-json-parser==0.2.1.1.post6
pathlib==1.0.1
pillow==11.3.0
pluggy==1.6.0
polyfactory==2.22.2
prometheus-fastapi-instrumentator==7.1.0
prometheus_client==0.22.1
propcache==0.3.2
protobuf==6.32.1
prov==2.1.1
psutil==7.0.0
puremagic==1.30
py-cpuinfo==9.0.0
pybase64==1.4.2
pyclipper==1.3.0.post6
pycountry==24.6.1
pycparser==2.23
pydantic==2.12.0a1
pydantic-extra-types==2.10.5
pydantic-settings==2.10.1
pydantic_core==2.37.2
pydot==4.0.1
Pygments==2.19.2
pylatexenc==2.10
PyMuPDF==1.26.4
pynvml==13.0.1
pyparsing==3.2.4
pypdfium2==4.30.0
pytest==8.4.2
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.1
python-json-logger==4.0.0rc1
python-multipart==0.0.20
python-pptx==1.0.2
pytorch-triton==3.4.0+git11ec6354
pytz==2025.2
pyxnat==1.6.3
PyYAML==6.0.2
pyzmq==27.1.0
ray==2.49.1
rdflib==7.1.4
referencing==0.36.2
regex==2025.9.1
requests==2.32.5
rich==14.1.0
rich-toolkit==0.15.1
rignore==0.6.4
rpds-py==0.27.1
rtree==1.4.1
safetensors==0.6.2
scikit-image==0.25.2
scipy==1.16.2
seaborn==0.13.2
semchunk==2.2.2
sentencepiece==0.2.1
sentry-sdk==3.0.0a6
setproctitle==1.3.7
setuptools==79.0.1
shapely==2.1.1
shellingham==1.5.4
simplejson==3.20.1
six==1.17.0
sniffio==1.3.1
soundfile==0.13.1
soupsieve==2.8
soxr==1.0.0
sse-starlette==3.0.2
starlette==0.47.3
structlog==25.4.0
sympy==1.14.0
tabulate==0.9.0
tenacity==9.1.2
termcolor==3.1.0
tifffile==2025.9.9
tiktoken==0.11.0
tokenizers==0.22.0
torch==2.9.0.dev20250804+cu128
torchaudio==2.8.0.dev20250804+cu128
torchvision==0.24.0.dev20250804+cu128
tqdm==4.67.1
traits==7.0.2
transformers==4.56.1
triton==3.4.0+git663e04e8
triton_kernels==1.0.0
typer==0.16.1
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
uvloop==0.21.0
vllm==0.10.1+gptoss
watchfiles==1.1.0
websockets==15.0.1
xgrammar==0.1.21
xlsxwriter==3.2.9
yarl==1.20.1
zipp==3.23.0
